test 멸망

+++ src/main/java/web/mvc/dto/AuctionDTO.java
@@ -21,5 +21,5 @@ public class AuctionDTO {
 
     private int status;
 
-    private int stockSeq;
+    private Long stockSeq;
 }
